# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/opening": {
      "get": {
        "description": "Show a job opening",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Openings"],
        "summary": "Show opening",
        "parameters": [
          {
            "type": "string",
            "description": "Opening identification",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handler.ShowOpeningResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          }
        }
      },
      "put": {
        "description": "Update a job opening",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Openings"],
        "summary": "Update opening",
        "parameters": [
          {
            "type": "string",
            "description": "Opening Identification",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "description": "Opening data to Update",
            "name": "opening",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handler.UpdateOpeningRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handler.UpdateOpeningResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Create a new job opening",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Openings"],
        "summary": "Create opening",
        "parameters": [
          {
            "description": "Request body",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handler.CreateOpeningRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handler.CreateOpeningResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete a new job opening",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Openings"],
        "summary": "Delete opening",
        "parameters": [
          {
            "type": "string",
            "description": "Opening identification",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handler.DeleteOpeningResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          }
        }
      }
    },
    "/openings": {
      "get": {
        "description": "List all job openings",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Openings"],
        "summary": "List openings",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handler.ListOpeningsResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "handler.CreateOpeningRequest": {
      "type": "object",
      "properties": {
        "company": {
          "type": "string"
        },
        "link": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "remote": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "salary": {
          "type": "integer"
        }
      }
    },
    "handler.CreateOpeningResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/schemas.OpeningResponse"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "handler.DeleteOpeningResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/schemas.OpeningResponse"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "handler.ErrorResponse": {
      "type": "object",
      "properties": {
        "errorCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "handler.ListOpeningsResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/schemas.OpeningResponse"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "handler.ShowOpeningResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/schemas.OpeningResponse"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "handler.UpdateOpeningRequest": {
      "type": "object",
      "properties": {
        "company": {
          "type": "string"
        },
        "link": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "remote": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "salary": {
          "type": "integer"
        }
      }
    },
    "handler.UpdateOpeningResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/schemas.OpeningResponse"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "schemas.OpeningResponse": {
      "type": "object",
      "properties": {
        "company": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "deteledAt": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "link": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "remote": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "salary": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    }
  }
}